workers/worker-template-jaeger/requirements-v2.txt
alembic==1.13.3
aniso8601==9.0.1
anyio==4.6.0
asn1crypto==1.5.1
asn1tools==0.167.0
async-property==0.2.2
async-timeout==5.0.0
attrs==24.2.0
Authlib==1.3.2
backports.tarfile==1.2.0
bcrypt==4.2.0
ber-tlv==0.0.6
bidict==0.23.1
bitstruct==8.19.0
blinker==1.8.2
build==1.2.2
cached-property==1.5.2
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
decorator==5.1.1
deepmerge==2.0
Deprecated==1.2.14
deprecation==2.1.0
docutils==0.21.2
ecdsa==0.19.0
ECPy==1.2.5
elastic-transport==8.15.0
elasticsearch==8.14.0
exceptiongroup==1.2.2
executing==2.1.0
Faker==30.0.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
flask-restx==1.3.0
Flask-SocketIO==5.3.7
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
Flask-WTF==1.2.1
googleapis-common-protos==1.59.1
greenlet==3.0.3
grpcio==1.66.2
h11==0.14.0
hatchling==1.25.0
httpcore==1.0.5
httpx==0.26.0
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
importlib_metadata==8.4.0
importlib_resources==6.4.5
isodate==0.6.1
iteration_utilities==0.12.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.8.0
Jinja2==3.1.4
jsonpath-ng==1.6.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jwcrypto==1.5.6
kafka==1.3.5
kafka-python==2.0.2
kafka-python-ng==2.2.2
keycloak==3.1.5
keyring==25.4.1
lxml==5.3.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
marshmallow-sqlalchemy==1.1.0
mdurl==0.1.2
more-itertools==10.5.0
moviepy==2.1.1
mrz==0.6.2
nh3==0.2.18
numpy==2.1.3
opentelemetry-api==1.28.2
opentelemetry-distro==0.48b0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-jaeger-proto-grpc==1.21.0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-flask==0.48b0
opentelemetry-instrumentation-kafka-python==0.48b0
opentelemetry-instrumentation-requests==0.48b0
opentelemetry-instrumentation-sqlalchemy==0.48b0
opentelemetry-instrumentation-wsgi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
opentelemetry-util-http==0.48b0
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
pillow==10.4.0
pkginfo==1.10.0
platformdirs==4.3.6
pluggy==1.5.0
ply==3.11
proglog==0.1.10
protobuf==4.25.5
psycopg2==2.9.9
pyasn1==0.6.1
pycparser==2.22
pycryptodome==3.20.0
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.9.0
PyMySQL==1.1.1
pyOpenSSL==24.2.1
pyparsing==3.1.4
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-jose==3.3.0
python-keycloak==4.5.0
python-socketio==5.11.4
pytz==2024.2
PyYAML==6.0.2
readme_renderer==44.0
redis==5.2.0
referencing==0.35.1
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.8.1
rpds-py==0.20.0
rsa==4.9
SecretStorage==3.3.3
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
spyne==2.14.0
SQLAlchemy==2.0.32
sqlalchemy-views==0.3.2
thrift==0.21.0
tomli==2.0.1
tqdm==4.67.1
trove-classifiers==2024.9.12
twine==5.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.2
varname==0.13.3
Werkzeug==3.0.4
wrapt==1.16.0
wsproto==1.2.0
WTForms==3.1.2
xmltodict==0.13.0
yt-dlp==2024.11.18
zeep==4.2.1
zipp==3.20.2
